Extract QueryColumn
