+++ server/web-service/src/test/java/org/bithon/server/web/service/datasource/api/GeneralQueryRequestTest.java
@@ -14,12 +14,11 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.web.service.api;
+package org.bithon.server.web.service.datasource.api;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.bithon.server.web.service.datasource.api.GeneralQueryRequest;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -45,9 +44,9 @@ public class GeneralQueryRequestTest {
 
         GeneralQueryRequest request = new ObjectMapper().configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
                                                         .readValue(json, GeneralQueryRequest.class);
-        Assert.assertEquals(GeneralQueryRequest.QueryColumn.class, request.getColumns().get(0).getClass());
-        Assert.assertEquals(GeneralQueryRequest.QueryColumn.class, request.getColumns().get(1).getClass());
-        Assert.assertEquals(GeneralQueryRequest.ExpressionQueryColumn.class, request.getColumns().get(2).getClass());
+        Assert.assertEquals(QueryColumn.DefaultQueryColumn.class, request.getColumns().get(0).getClass());
+        Assert.assertEquals(QueryColumn.DefaultQueryColumn.class, request.getColumns().get(1).getClass());
+        Assert.assertEquals(QueryColumn.ExpressionQueryColumn.class, request.getColumns().get(2).getClass());
 
         Assert.assertEquals("instanceUpTime", request.getOrderBy().getName());
         Assert.assertEquals("desc", request.getOrderBy().getOrder());
